+++ SpaceInvadersRemake/SpaceInvadersRemake/SpaceInvadersRemake/Controller/WaveAI.cs
@@ -23,7 +23,7 @@ namespace SpaceInvadersRemake.Controller
         /// </remarks>
         /// <param name="shootingFrequencyMultiplier">Die Schussfrequenz.</param>
         /// <param name="controllees">Die GameItem, die der Controller kontrollieren soll.</param>
-      protected  WaveAI(int shootingFrequency, ICollection<IGameItem> controllees,Vector2 velocityIncrease)
+      protected  WaveAI(float shootingFrequency, ICollection<IGameItem> controllees,Vector2 velocityIncrease)
             : base(shootingFrequency, null, velocityIncrease) //null wird das single controllee gesetzt 
         {
